Pull out xml utils into mixin.
