+++ app/models/pb_core.rb
@@ -4,15 +4,12 @@ require 'solrizer'
 require_relative '../../lib/aapb'
 require_relative 'exhibit'
 require_relative '../../lib/html_scrubber'
-
+require_relative 'xml_backed'
 require_relative 'organization'
 
 class PBCore # rubocop:disable Metrics/ClassLength
   # rubocop:disable Style/EmptyLineBetweenDefs
-  def initialize(xml)
-    @xml = xml
-    @doc = REXML::Document.new xml
-  end
+  include XmlBacked
   def descriptions
     @descriptions ||= xpaths('/*/pbcoreDescription').map { |description| HtmlScrubber.scrub(description) }
   end
@@ -309,51 +306,6 @@ class PBCore # rubocop:disable Metrics/ClassLength
   end
 
   private
-
-  class NoMatchError < StandardError
-  end
-
-  def xpath(xpath)
-    REXML::XPath.match(@doc, xpath).tap do |matches|
-      if matches.length != 1
-        fail NoMatchError, "Expected 1 match for '#{xpath}'; got #{matches.length}"
-      else
-        return PBCore.text_from(matches.first)
-      end
-    end
-  end
-
-  def xpaths(xpath)
-    REXML::XPath.match(@doc, xpath).map { |node| PBCore.text_from(node) }
-  end
-
-  def self.text_from(node)
-    ((node.respond_to?('text') ? node.text : node.value) || '').strip
-  end
-
-  def pairs_by_type(element_xpath, attribute_xpath)
-    REXML::XPath.match(@doc, element_xpath).map do |node|
-      key = REXML::XPath.first(node, attribute_xpath)
-      [
-        key ? key.value : nil,
-        node.text
-      ]
-    end
-  end
-
-  def hash_by_type(element_xpath, attribute_xpath)
-    Hash[pairs_by_type(element_xpath, attribute_xpath)]
-  end
-
-# TODO: If we can just iterate over pairs, we don't need either of these.
-#
-#  def multi_hash_by_type(element_xpath, attribute_xpath) # Not tested
-#    Hash[
-#      pairs_by_type(element_xpath, attribute_xpath).group_by{|(key,value)| key}.map{|key,pair_list|
-#        [key, pair_list.map{|(key,value)| value}]
-#      }
-#    ]
-#  end
 
   # These methods are only used by to_solr.
